+++ src/html_element/types.ts
@@ -1,14 +1,9 @@
 import type { ChangeEvent, HTMLInputTypeAttribute } from "react";
+import type { StoreBindingMeta } from "../binding/use_store_binding";
 
 export type InputDisplayValue = string | number | readonly string[];
 export type InputStateValue = string | string[];
-export type InputControlValue =
-  | string
-  | readonly string[]
-  | number
-  | boolean
-  | FileList
-  | null;
+export type InputControlValue = string | readonly string[] | number | boolean | FileList | null;
 
 export type StoreInputOptions<TInputElement> = {
   type?: HTMLInputTypeAttribute;
@@ -20,9 +15,8 @@ export type StoreInputOptions<TInputElement> = {
   onChange?: (event: ChangeEvent<TInputElement>) => void;
 };
 
-export type StoreInputBindingOptions<TInputElement, TValue> =
-  StoreInputOptions<TInputElement> & {
-    resetValue?: TValue;
+export type StoreInputBindingOptions<TInputElement, TValue> = StoreInputOptions<TInputElement> & {
+  resetValue?: TValue;
 };
 
 export type StoreInputDomProps<TInputElement> = {
@@ -31,9 +25,7 @@ export type StoreInputDomProps<TInputElement> = {
   onChange: (event: ChangeEvent<TInputElement>) => void;
 };
 
-export type StoreInputMeta<TError> =
-  | { valid: true; error?: never }
-  | { valid: false; error: TError };
+export type StoreInputMeta<TError> = StoreBindingMeta<TError>;
 
 export type BindingInputResult<TInputElement, TError> = {
   inputProps: StoreInputDomProps<TInputElement>;